evaluate tab changes
